Extract focus-safe DOM swap into sections/dom-swap.js
First region extraction of P-09 cut 2. focusSafeSwap is the primitive both _reloadInPlace and reloadNode delegate to — extracting it out of dashboard-main.js is a precondition for eventually extracting both callers. No new cross-module imports beyond what's already routed through sections/reload-guards.js (_defaultReloadSkipIf) and sections/local-subtree-reload.js (_capture/_restoreDetailsOpenState). 101 LOC moved.
